# perf: create DocumentArray index map lazily