--- a/tests/unit/types/arrays/documentarray/test_documentarray.py
+++ b/tests/unit/types/arrays/documentarray/test_documentarray.py
@@ -347,9 +347,10 @@ def test_none_extend():
 
 
 def test_lazy_index_map():
-    da = DocumentArray([Document() for _ in range(100)])
+    da = DocumentArray([Document(id=str(i), text=f'document_{i}') for i in range(100)])
     assert da._id_to_index is None
 
     # build index map
-    assert len(da._index_map.keys()) == 100
+    assert da['0'].text == 'document_0'
     assert da._id_to_index is not None
+    assert len(da._index_map.keys()) == 100