closes https://github.com/jina-ai/internal-tasks/issues/279

results:
```python
from jina import Document
from jina.logging.profile import TimeContext

doc = Document()
with TimeContext('appending chunks'):
    for _ in range(10000):
        doc.chunks.append(
            Document(
                text='hey',
                offset=0,
                weight=1.0 ,
                location=[1, 2],
            )
        )
```

on master:
```
appending chunks ...	appending chunks takes 31 seconds (31.40s)
```

on feature branch:
```
appending chunks ...	appending chunks takes 0 seconds (0.64s)
```