@alaeddine-13 2021-11-18
> Add a test in the types than make sure that a DocumentArray has `_id_to_index` to None at the beginning?

added